standard setup

+++ src/DWIS.Docker/DWIS.Docker.Clients/DWISModulesConfigurationClient.cs
@@ -1,4 +1,5 @@
 ﻿using DWIS.Docker.Constants;
+using DWIS.Docker.Models;
 
 namespace DWIS.Docker.Clients
 {
@@ -8,7 +9,7 @@ namespace DWIS.Docker.Clients
 
 
 
-        public ComposerConfig? GetComposerConfigFromFile()
+        public DWIS.AdviceComposer.Service.Configuration? GetComposerConfigFromFile()
         {
             string filePath = Path.Combine(ImageNames.COMPOSER_WINDOWS_LOCALPATH, ImageNames.COMPOSER_CONFIGFILENAME);
 
@@ -19,11 +20,11 @@ namespace DWIS.Docker.Clients
             else
             {
                 string json = System.IO.File.ReadAllText(filePath);
-                var config = System.Text.Json.JsonSerializer.Deserialize<ComposerConfig>(json);
+                var config = System.Text.Json.JsonSerializer.Deserialize<DWIS.AdviceComposer.Service.Configuration>(json);
                 return config;
             }
         }
-        
+
         public ConfigType? GetConfigurationFromFile<ConfigType>(string localPath, string fileName) where ConfigType : class
         {
             string filePath = Path.Combine(localPath, fileName);
@@ -55,19 +56,16 @@ namespace DWIS.Docker.Clients
             {
                 Directory.CreateDirectory(localPath);
             }
-            System.IO.File.WriteAllText(Path.Combine(localPath, fileName), config);
+            if(!File.Exists(Path.Combine(localPath, fileName)))
+            {
+                System.IO.File.WriteAllText(Path.Combine(localPath, fileName), config);
+            }
         }
 
-
-
-        public class ComposerConfig
+        public void SaveStandardConfigToFile(StandardSetUpItem item)
         {
-            public TimeSpan LoopDuration { get; set; } = TimeSpan.FromSeconds(1.0);
-            public string? OPCUAURL { get; set; } = "opc.tcp://localhost:48030";
-            public TimeSpan ControllerObsolescence { get; set; } = TimeSpan.FromSeconds(5.0);
-            public TimeSpan ProcedureObsolescence { get; set; } = TimeSpan.FromSeconds(5.0);
-            public TimeSpan FaultDetectionIsolationAndRecoveryObsolescence { get; set; } = TimeSpan.FromSeconds(5.0);
-            public TimeSpan SafeOperatingEnvelopeObsolescence { get; set; } = TimeSpan.FromSeconds(5.0);
+            if(item.ConfigurationRequired)
+                SaveConfigToFile(item.DefaultConfigContent, item.ConfigLocalPath, item.ConfigFileName);
         }
     }
 }
